requirements.txt
pyyaml
pexpect
gpiod
